+++ NEWS.md
@@ -1,5 +1,10 @@
 # aRt 
 
+# aRt 1.1.8
+
+* add `sunbursts()` function
+* add panel function
+
 # aRt 1.1.7
 
 * add GitHub actions